* Updated the model Metar with new Pressure datas and deleted Altimeter reference * created new function to convert pressure datas in Metar
