+++ metar_taf_parser/commons/converter.py
@@ -1,3 +1,6 @@
+from metar_taf_parser.model.model import Pressure
+
+
 def degrees_to_cardinal(input):
     if input.isnumeric():
         degrees = int(input)
@@ -21,5 +24,20 @@ def convert_temperature(input):
     return int(input)
 
 
-def convert_inches_mercury_to_pascal(input):
-    return 33.8639 * input
+def convert_inhg_to_hpa(input):
+    return '{:5.2f}'.format(float(input) * 0.029529983071445)
+
+
+def convert_hpa_to_inhg(input):
+    return '{:4.0f}'.format(float(input) / 0.029529983071445)
+
+
+def convert_pressure(input):
+    pressure = Pressure()
+    if input.startswith('Q'):
+        pressure.pressure = '{:4.0f}'.format(float(input.split('Q')[1]))
+        pressure.unit = 'hPa'
+    else:
+        pressure.pressure = '{:5.2f}'.format(float(input.split('A')[1]) / 100)
+        pressure.unit = 'inHg'
+    return pressure
